requirements.txt
asgiref==3.3.0
Django==3.1.3
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-taggit==1.3.0
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
